moved serialization mocks to separate file
