+++ test/unit/serialization/test_get_class.py
@@ -4,7 +4,8 @@ import pytest
 from fedot.core.serializers.json_helpers import DELIMITER, _get_class
 
 from .dataclasses.serialization_dataclasses import GetClassCase
-from .test_input import Bar, Baz, MockGraph, MockNode, MockOperation, TestEnum, foo
+from .mocks.serialization_mocks import MockGraph, MockNode, MockOperation
+from .test_input import Bar, Baz, TestEnum, foo
 
 GET_CLASS_CASES = [
     GetClassCase(
